Extract piecewise_poly_path declaration to a separate header
