+++ cpp/tests/test_algorithm.cpp
@@ -3,6 +3,7 @@
 #include <toppra/constraint.hpp>
 #include <toppra/constraint/linear_joint_acceleration.hpp>
 #include <toppra/constraint/linear_joint_velocity.hpp>
+#include <toppra/geometric_path/piecewise_poly_path.hpp>
 #include <toppra/geometric_path.hpp>
 #include <toppra/toppra.hpp>
 #include "toppra/algorithm.hpp"
