+++ cpp/tests/test_constraints.cpp
@@ -6,6 +6,7 @@
 #endif
 
 #include <toppra/geometric_path.hpp>
+#include <toppra/geometric_path/piecewise_poly_path.hpp>
 
 #include "gtest/gtest.h"
